+++ packages/react-components/react-combobox/src/components/Option/useOption.tsx
@@ -1,10 +1,8 @@
 import * as React from 'react';
 import { getIntrinsicElementProps, useId, useMergedRefs, slot } from '@fluentui/react-utilities';
-import { useContextSelector } from '@fluentui/react-context-selector';
 import { useActiveDescendantContext } from '@fluentui/react-aria';
 import { CheckmarkFilled, Checkmark12Filled } from '@fluentui/react-icons';
-import { ComboboxContext } from '../../contexts/ComboboxContext';
-import { ListboxContext } from '../../contexts/ListboxContext';
+import { useListboxContext_unstable } from '../../contexts/ListboxContext';
 import type { OptionValue } from '../../utils/OptionCollection.types';
 import type { OptionProps, OptionState } from './Option.types';
 
@@ -56,18 +54,17 @@ export const useOption_unstable = (props: OptionProps, ref: React.Ref<HTMLElemen
     [id, disabled, optionText, optionValue],
   );
 
-  const { controller: activeDescendantController } = useActiveDescendantContext();
-
   // context values
-  const multiselect = useContextSelector(ListboxContext, ctx => ctx.multiselect);
-  const registerOption = useContextSelector(ListboxContext, ctx => ctx.registerOption);
-  const selected = useContextSelector(ListboxContext, ctx => {
+  const { controller: activeDescendantController } = useActiveDescendantContext();
+  const multiselect = useListboxContext_unstable(ctx => ctx.multiselect);
+  const registerOption = useListboxContext_unstable(ctx => ctx.registerOption);
+  const selected = useListboxContext_unstable(ctx => {
     const selectedOptions = ctx.selectedOptions;
 
     return !!optionValue && !!selectedOptions.find(o => o === optionValue);
   });
-  const selectOption = useContextSelector(ListboxContext, ctx => ctx.selectOption);
-  const setOpen = useContextSelector(ComboboxContext, ctx => ctx.setOpen);
+  const selectOption = useListboxContext_unstable(ctx => ctx.selectOption);
+  const onOptionClick = useListboxContext_unstable(ctx => ctx.onOptionClick);
 
   // check icon
   let CheckIcon: React.ReactNode = <CheckmarkFilled />;
@@ -81,17 +78,12 @@ export const useOption_unstable = (props: OptionProps, ref: React.Ref<HTMLElemen
       return;
     }
 
-    // clicked option should always become active option
     activeDescendantController.focus(id);
-
-    // close on option click for single-select options in a combobox
-    if (!multiselect) {
-      setOpen?.(event, false);
-    }
 
     // handle selection change
     selectOption(event, optionData);
 
+    onOptionClick(event);
     props.onClick?.(event);
   };