+++ packages/react-components/react-teaching-popover-preview/src/components/TeachingPopoverCarousel/TeachingPopoverCarousel.tsx
@@ -5,6 +5,7 @@ import { useTeachingPopoverCarouselStyles_unstable } from './useTeachingPopoverC
 import { useCustomStyleHook_unstable } from '@fluentui/react-shared-contexts';
 import type { TeachingPopoverCarouselProps } from './TeachingPopoverCarousel.types';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
+import { useTeachingPopoverCarouselContextValues_unstable } from './TeachingPopoverCarouselContext';
 
 /**
  * Define a styled TeachingPopoverCarousel, using the `useTeachingPopoverCarousel_unstable` and `useTeachingPopoverCarouselStyles_unstable`
@@ -18,7 +19,9 @@ export const TeachingPopoverCarousel: ForwardRefComponent<TeachingPopoverCarouse
 
     useCustomStyleHook_unstable('useTeachingPopoverCarouselStyles_unstable')(state);
 
-    return renderTeachingPopoverCarousel_unstable(state);
+    const contextValues = useTeachingPopoverCarouselContextValues_unstable(state);
+
+    return renderTeachingPopoverCarousel_unstable(state, contextValues);
   },
 );